+++ tests/test_bitarray.cc
@@ -17,7 +17,7 @@ using ::libk2tree::uchar;
 
 TEST(BitArrayChar, Constructor) {
   BitArray<uchar> bs(1000);
-  for (int i = 0; i < 1000; ++i)
+  for (uint i = 0; i < 1000; ++i)
     ASSERT_EQ(0, bs.GetBit(i));
 }
 TEST(BitArray, SetBit0) {
@@ -26,13 +26,13 @@ TEST(BitArray, SetBit0) {
   ASSERT_EQ(1, bs.GetBit(0));
 }
 TEST(BitArrayChar, SetBit) {
-  srand(time(NULL));
-  for (int i = 0; i < 100; ++i) {
+  srand((uint) time(NULL));
+  for (uint i = 0; i < 100; ++i) {
     size_t N = rand()%10000 + 1000;
     BitArray<uchar> bs(N);
 
     for (int j = 0; j < 100 ; ++j) {
-      size_t pos = rand()%N;
+      size_t pos = (uint) rand()%N;
       bs.SetBit(pos);
       ASSERT_EQ(1, bs.GetBit(pos));
     }
@@ -40,13 +40,13 @@ TEST(BitArrayChar, SetBit) {
 }
 
 TEST(BitArrayChar, CleanBit) {
-  srand(time(NULL));
+  srand((uint) time(NULL));
   for (int i = 0; i < 100; ++i) {
     size_t N = rand()%10000 + 1000;
     BitArray<uchar> bs(N);
 
     for (int j = 0; j < 100 ; ++j) {
-      size_t pos = rand()%N;
+      size_t pos = (uint) rand()%N;
       bs.CleanBit(pos);
       ASSERT_EQ(0, bs.GetBit(pos));
     }
@@ -57,17 +57,17 @@ TEST(BitArrayChar, CleanBit) {
 TEST(BitArrayInt, Constructor) {
   BitArray<uint> bs(1000);
 
-  for (int i = 0; i < 1000; ++i)
+  for (uint i = 0; i < 1000; ++i)
     ASSERT_EQ(0, bs.GetBit(i));
 }
 TEST(BitArrayInt, SetBit) {
-  srand(time(NULL));
+  srand((uint) time(NULL));
   for (int i = 0; i < 100; ++i) {
     size_t N = rand()%10000 + 1000;
     BitArray<uint> bs(N);
 
     for (int j = 0; j < 100 ; ++j) {
-      size_t pos = rand()%N;
+      size_t pos = (uint) rand()%N;
       bs.SetBit(pos);
       ASSERT_EQ(1, bs.GetBit(pos));
     }
@@ -76,31 +76,31 @@ TEST(BitArrayInt, SetBit) {
 
 
 TEST(BitArrayInt, CleanBit) {
-  srand(time(NULL));
+  srand((uint) time(NULL));
   for (int i = 0; i < 100; ++i) {
     size_t N = rand()%10000 + 1000;
     BitArray<uint> bs(N);
 
     for (int j = 0; j < 100 ; ++j) {
-      size_t pos = rand()%N;
+      size_t pos = (uint) rand()%N;
       bs.CleanBit(pos);
       ASSERT_EQ(0, bs.GetBit(pos));
     }
   }
 }
 TEST(BitArray, CopyConstructor) {
-  srand(time(NULL));
+  srand((uint) time(NULL));
   for (int i = 0; i < 100; ++i) {
     size_t N = rand()%10000 + 1000;
     BitArray<uint> bs(N);
 
     for (int j = 0; j < 100 ; ++j) {
-      size_t pos = rand()%N;
+      size_t pos = (uint) rand()%N;
       bs.SetBit(pos);
     }
     BitArray<uint> bs2(bs);
     for (int j = 0; j < 100 ; ++j) {
-      size_t pos = rand()%N;
+      size_t pos = (uint) rand()%N;
       ASSERT_EQ(bs.GetBit(pos), bs2.GetBit(pos));
     }
   }
