+++ tests/test_k2treepartition.cc
@@ -27,26 +27,26 @@ typedef unsigned int uint;
 
 shared_ptr<K2TreePartition> BuildPartition(vector<vector<bool>> *matrix) {
   string filename = "partition_test";
-  int n = 1000;
+  uint n = 1000;
 
-  int e = rand()%(n*10) + 1;
+  uint e = (uint) rand()%(n*10) + 1;
 
   matrix->resize(n, vector<bool>(n, false));
-  for (int i = 0; i < e; ++i) {
-    int p = rand()%n;
-    int q = rand()%n;
+  for (uint i = 0; i < e; ++i) {
+    uint p = (uint) rand()%n;
+    uint q = (uint) rand()%n;
     (*matrix)[p][q] = true;
   }
 
-  int k0 = 10;
-  int subm = n/k0;
+  uint k0 = 10;
+  uint subm = n/k0;
 
 
   K2TreePartitionBuilder b(n, subm, 4, 2, 2, 3, filename);
-  for (int row = 0; row < k0; ++row) {
-    for (int col = 0; col < k0; ++col) {
-      for (int i = 0; i < subm; ++i) {
-        for (int j = 0; j < subm; ++j) {
+  for (uint row = 0; row < k0; ++row) {
+    for (uint col = 0; col < k0; ++col) {
+      for (uint i = 0; i < subm; ++i) {
+        for (uint j = 0; j < subm; ++j) {
           if ((*matrix)[i + row*subm][j + col*subm])
             b.AddLink(i + row*subm, j + col*subm);
         }
@@ -62,7 +62,7 @@ shared_ptr<K2TreePartition> BuildPartition(vector<vector<bool>> *matrix) {
 }
 
 TEST(k2treepartition, CheckLink) {
-  srand(time(NULL));
+  srand((uint) time(NULL));
   vector<vector<bool>> matrix;
   shared_ptr<K2TreePartition> tree = BuildPartition(&matrix);
